Imports simulator/events.py code in relevant EventClass methods
